review logic

+++ app/Http/Controllers/ReviewController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\CreateReviewRequest;
 use App\Http\Requests\UpdateReviewsRequest;
+use App\Http\Responses\ApiErrorResponse;
 use App\Models\Review;
 use App\Models\Product;
 use Illuminate\Http\Response;
@@ -14,10 +15,19 @@ class ReviewController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Product $product)
     {
+        $reviews = $product->reviews()->get();
+        if ($reviews->isEmpty()) {
+            return new ApiErrorResponse(
+                "This product has not been reviewed",
+                null,
+                Response::HTTP_NO_CONTENT
+            );
+        }
+
         return new ApiSuccessResponse(
-            Review::all(),
+            $reviews,
             ['message' => "List of all Reviews"]
         );
     }
@@ -34,25 +44,40 @@ class ReviewController extends Controller
             ["message" => "Review Created Successfully"],
             Response::HTTP_CREATED
         );
-
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Review $review)
+    public function show(Product $product, Review $review)
     {
+        if ($review->product_id !== $product->id) {
+            return new ApiErrorResponse(
+                "The review does not belong to this product",
+                null,
+                Response::HTTP_NOT_FOUND
+            );
+        }
+
         return new ApiSuccessResponse(
             $review,
-            ["message" => "Single Category"]
+            ["message" => "Single Review"]
         );
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateReviewsRequest $request, Review $review)
+    public function update(UpdateReviewsRequest $request, Product $product, Review $review)
     {
+        if ($review->product_id !== $product->id) {
+            return new ApiErrorResponse(
+                "The review does not belong to this product",
+                null,
+                Response::HTTP_NOT_FOUND
+            );
+        }
+
         $review->update($request->all());
 
         return new ApiSuccessResponse(
@@ -64,8 +89,16 @@ class ReviewController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Review $review)
+    public function destroy(Product $product, Review $review)
     {
+        if ($review->product_id !== $product->id) {
+            return new ApiErrorResponse(
+                "The review does not belong to this product",
+                null,
+                Response::HTTP_NOT_FOUND
+            );
+        }
+
         $review->delete();
 
         return new ApiSuccessResponse(
@@ -73,7 +106,5 @@ class ReviewController extends Controller
             ['message' => "Product Deleted Successfully"],
             Response::HTTP_NO_CONTENT
         );
-
-
     }
 }
